Change rate stats classes to use uint64_t
This avoids integer overflows when size_t is 32 bits.

Bug: webrtc:11247

+++ rtc_base/rate_statistics.cc
@@ -54,7 +54,7 @@ void RateStatistics::Reset() {
     buckets_[i] = Bucket();
 }
 
-void RateStatistics::Update(size_t count, int64_t now_ms) {
+void RateStatistics::Update(uint64_t count, int64_t now_ms) {
   if (now_ms < oldest_time_) {
     // Too old data is ignored.
     return;
@@ -77,7 +77,7 @@ void RateStatistics::Update(size_t count, int64_t now_ms) {
   ++num_samples_;
 }
 
-absl::optional<uint32_t> RateStatistics::Rate(int64_t now_ms) const {
+absl::optional<uint64_t> RateStatistics::Rate(int64_t now_ms) const {
   // Yeah, this const_cast ain't pretty, but the alternative is to declare most
   // of the members as mutable...
   const_cast<RateStatistics*>(this)->EraseOld(now_ms);
@@ -91,7 +91,7 @@ absl::optional<uint32_t> RateStatistics::Rate(int64_t now_ms) const {
   }
 
   float scale = scale_ / active_window_size;
-  return static_cast<uint32_t>(accumulated_count_ * scale + 0.5f);
+  return static_cast<uint64_t>(accumulated_count_ * scale + 0.5f);
 }
 
 void RateStatistics::EraseOld(int64_t now_ms) {
